refactor(core): extract categorical palettes from scales/train.ts
Move pure categorical palette tables and CATEGORICAL_SCHEMES into
categorical-palettes.ts so scheme data is not mixed with scale training.
Public ./scales/train.js re-exports preserve the same ES bindings for
index, editions, and identity === checks.
